+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -118,7 +118,7 @@ class DeploymentsTools(BaseModel):
     id: Optional[float] = None
 
 
-ModelType = Literal[
+DeploymentsModelType = Literal[
     "chat",
     "completion",
     "embedding",
@@ -151,13 +151,13 @@ class DeploymentsResponseFormat2(BaseModel):
 DeploymentsResponseFormatDeploymentsType = Literal["json_schema"]
 
 
-class ResponseFormatJSONSchemaTypedDict(TypedDict):
+class DeploymentsResponseFormatJSONSchemaTypedDict(TypedDict):
     name: str
     strict: bool
     schema_: Dict[str, Any]
 
 
-class ResponseFormatJSONSchema(BaseModel):
+class DeploymentsResponseFormatJSONSchema(BaseModel):
     name: str
 
     strict: bool
@@ -167,13 +167,13 @@ class ResponseFormatJSONSchema(BaseModel):
 
 class DeploymentsResponseFormat1TypedDict(TypedDict):
     type: DeploymentsResponseFormatDeploymentsType
-    json_schema: ResponseFormatJSONSchemaTypedDict
+    json_schema: DeploymentsResponseFormatJSONSchemaTypedDict
 
 
 class DeploymentsResponseFormat1(BaseModel):
     type: DeploymentsResponseFormatDeploymentsType
 
-    json_schema: ResponseFormatJSONSchema
+    json_schema: DeploymentsResponseFormatJSONSchema
 
 
 DeploymentsResponseFormatTypedDict = TypeAliasType(
@@ -211,7 +211,7 @@ DeploymentsEncodingFormat = Literal["float", "base64"]
 r"""The format to return the embeddings"""
 
 
-class ModelParametersTypedDict(TypedDict):
+class DeploymentsModelParametersTypedDict(TypedDict):
     r"""Model Parameters: Not all parameters apply to every model"""
 
     temperature: NotRequired[float]
@@ -253,7 +253,7 @@ class ModelParametersTypedDict(TypedDict):
     r"""The format to return the embeddings"""
 
 
-class ModelParameters(BaseModel):
+class DeploymentsModelParameters(BaseModel):
     r"""Model Parameters: Not all parameters apply to every model"""
 
     temperature: Optional[float] = None
@@ -528,26 +528,26 @@ class DeploymentsMessages(BaseModel):
     tool_calls: Optional[List[DeploymentsDeploymentsToolCalls]] = None
 
 
-class PromptConfigTypedDict(TypedDict):
+class DeploymentsPromptConfigTypedDict(TypedDict):
     tools: List[DeploymentsToolsTypedDict]
     model: str
-    model_type: ModelType
+    model_type: DeploymentsModelType
     r"""The type of the model"""
-    model_parameters: ModelParametersTypedDict
+    model_parameters: DeploymentsModelParametersTypedDict
     r"""Model Parameters: Not all parameters apply to every model"""
     provider: DeploymentsProvider
     messages: List[DeploymentsMessagesTypedDict]
 
 
-class PromptConfig(BaseModel):
+class DeploymentsPromptConfig(BaseModel):
     tools: List[DeploymentsTools]
 
     model: str
 
-    model_type: ModelType
+    model_type: DeploymentsModelType
     r"""The type of the model"""
 
-    model_parameters: ModelParameters
+    model_parameters: DeploymentsModelParameters
     r"""Model Parameters: Not all parameters apply to every model"""
 
     provider: DeploymentsProvider
@@ -566,7 +566,7 @@ class DataTypedDict(TypedDict):
     r"""The deployment unique key"""
     description: str
     r"""An arbitrary string attached to the object. Often useful for displaying to users."""
-    prompt_config: PromptConfigTypedDict
+    prompt_config: DeploymentsPromptConfigTypedDict
     version: str
     r"""THe version of the deployment"""
 
@@ -587,7 +587,7 @@ class Data(BaseModel):
     description: str
     r"""An arbitrary string attached to the object. Often useful for displaying to users."""
 
-    prompt_config: PromptConfig
+    prompt_config: DeploymentsPromptConfig
 
     version: str
     r"""THe version of the deployment"""