+++ packages/orq-rc/src/orq_ai_sdk/sdkconfiguration.py
@@ -26,9 +26,9 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "2.0"
-    sdk_version: str = "1.32.0-rc.5"
+    sdk_version: str = "1.32.0-rc.6"
     gen_version: str = "2.481.0"
-    user_agent: str = "speakeasy-sdk/python 1.32.0-rc.5 2.481.0 2.0 orq-ai-sdk"
+    user_agent: str = "speakeasy-sdk/python 1.32.0-rc.6 2.481.0 2.0 orq-ai-sdk"
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None